# docs/swagger.json
{
  "openapi": "3.1.0",
  "info": {
    "version": "1.0.0",
    "title": "AquaTrack",
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "description": "AquaTrack with user authorization"
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    },
    {
      "url": "https://aquatrack-duna-server.onrender.com"
    }
  ],
  "paths": {
...
    "/water": {
      "post": {
        "tags": [
          "Water"
        ],
        "summary": "Add Water",
        "operationId": "addWater",
        "description": "Create a note about portion of water",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "amount",
                  "time",
                  "percentage"
                ],
                "properties": {
                  "amount": {
                    "description": "Amount of a water portion",
                    "type": "integer",
                    "example": 250
                  },
                  "time": {
                    "description": "Time of drinking",
                    "type": "string",
                    "example": "2024-11-24T07:00:00"
                  },
                  "percentage": {
                    "description": "Part in daily norm, percentage",
                    "type": "integer",
                    "example": 20
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Creates a water",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "status",
                    "message",
                    "data"
                  ],
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 201
                    },
                    "message": {
                      "type": "string",
                      "example": "Water added successfully"
                    },
                    "data": {
                      "type": "object",
                      "$ref": "#/components/schemas/water"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "404": {
            "$ref": "#/components/responses/404"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
    "schemas": {
...
      "water": {
        "type": "object",
        "required": [
          "amount",
          "time",
          "percentage",
          "userId"
        ],
        "properties": {
          "_id": {
            "description": "Id of water portion",
            "type": "string",
            "example": "674340b9d687f43970852f8f"
          },
          "amount": {
            "description": "Amount of a water portion",
            "type": "integer",
            "example": 250
          },
          "time": {
            "description": "Time of drinking",
            "type": "string",
            "example": "2024-11-24 07:00:00"
          },
          "percentage": {
            "description": "Part in daily norm, percentage",
            "type": "integer",
            "example": 20
          },
          "userId": {
            "description": "Id of user who drinks",
            "type": "string",
            "example": "664ca67erae7f10c88b598y"
          }
        }
      }
    },
    "responses": {
...
      "401": {
        "description": "Unauthorized error",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": [
                "status",
                "message",
                "data"
              ],
              "properties": {
                "status": {
                  "type": "integer",
                  "example": 401
                },
                "message": {
                  "type": "string",
                  "example": "UnauthorizedError"
                },
                "data": {
                  "type": "object",
                  "required": [
                    "message"
                  ],
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Token is expired or invalid."
                    }
                  }
                }
              }
            }
          }
        }
      },
      "404": {
        "description": "Not found",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": [
                "status",
                "message",
                "error"
              ],
              "properties": {
                "status": {
                  "type": "integer",
                  "example": 404
                },
                "message": {
                  "type": "string",
                  "example": "NotFoundError"
                },
                "error": {
                  "type": "object",
                  "required": [
                    "message"
                  ],
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "User not found"
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  }
}